Release: Beta 10
This changeset improves and cleans up the JavaScript library, with
the goal of easier integration, cleaner execution, and better
maintainability.

Running worklog:
- [x] Split context into multiple files
- [x] Version bump -> 'beta10'
- [x] Refactor and cleanup global state
  - [x] Add support for user ID
  - [x] Add support for order ID
