+++ .idea/jsLibraryMappings.xml
@@ -2,6 +2,5 @@
 <project version="4">
   <component name="JavaScriptLibraryMappings">
     <file url="PROJECT" libraries="{Bloombox Schema, Protobuf JS}" />
-    <excludedPredefinedLibrary name="JS/schema/node_modules" />
   </component>
 </project>
