+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@bloombox/js-client",
-  "version": "1.0.0-beta9",
+  "version": "1.0.0-beta10",
   "license": "Apache-2.0",
   "private": false,
   "description": "JavaScript API client for Bloombox",
