+++ src/license.txt
@@ -10,7 +10,7 @@
  *    | $$$$$$$| $|  $$$$$$|  $$$$$$| $$ | $$ | $| $$$$$$$|  $$$$$$//$$/\  $$      |  $$$$$$|  $$$$$$/
  *    |_______/|__/\______/ \______/|__/ |__/ |__|_______/ \______/|__/  \__/       \______/ \______/
  *
- * Bloombox JS v1.0.0-beta9
+ * Bloombox JS v1.0.0-beta10
  *
  * Copyright (c) 2017, Momentum Ideas, Co. All rights reserved. Distributed under the Apache License, Version 2.0,
  * last revised January 2004 and enclosed in this codebase in 'LICENSE.txt'. Also, see third-party license notices
