+++ src/telemetry/context/base.js
@@ -33,110 +33,31 @@ goog.require('bloombox.logging.log');
 
 goog.require('bloombox.menu.Section');
 
+goog.require('bloombox.telemetry.Collection');
+goog.require('bloombox.telemetry.buildBrowserContext');
+goog.require('bloombox.telemetry.buildNativeContext');
+goog.require('bloombox.telemetry.buildWebappContext');
+
 goog.require('bloombox.util.Exportable');
 goog.require('bloombox.util.Serializable');
-goog.require('bloombox.util.b64');
-goog.require('bloombox.util.generateUUID');
-
-goog.require('goog.labs.userAgent.device');
-goog.require('goog.userAgent');
-goog.require('goog.userAgent.platform');
-goog.require('goog.userAgent.product');
 
 goog.require('proto.analytics.Context');
 goog.require('proto.analytics.Scope');
 goog.require('proto.analytics.context.APIClient');
-goog.require('proto.analytics.context.ApplicationType');
 goog.require('proto.analytics.context.BrowserDeviceContext');
-goog.require('proto.analytics.context.Collection');
 goog.require('proto.analytics.context.DeviceApplication');
 goog.require('proto.analytics.context.DeviceLibrary');
-goog.require('proto.analytics.context.DeviceOS');
-goog.require('proto.analytics.context.DeviceRole');
-goog.require('proto.analytics.context.DeviceScreen');
 goog.require('proto.analytics.context.NativeDeviceContext');
-goog.require('proto.analytics.context.OSType');
-goog.require('proto.analytics.context.PixelSize');
-goog.require('proto.analytics.context.ScreenOrientation');
-goog.require('proto.analytics.context.WebApplication');
+
 goog.require('proto.commerce.OrderKey');
-goog.require('proto.device.DeviceType');
 goog.require('proto.identity.UserKey');
 goog.require('proto.partner.PartnerDeviceKey');
 goog.require('proto.partner.PartnerKey');
 goog.require('proto.partner.PartnerLocationKey');
 goog.require('proto.structs.VersionSpec');
 
-goog.provide('bloombox.telemetry.Collection');
 goog.provide('bloombox.telemetry.Context');
 goog.provide('bloombox.telemetry.ContextException');
-
-goog.provide('bloombox.telemetry.buildBrowserContext');
-goog.provide('bloombox.telemetry.buildNativeContext');
-goog.provide('bloombox.telemetry.globalContext');
-goog.provide('bloombox.telemetry.resolveFingerprint');
-goog.provide('bloombox.telemetry.resolveSessionID');
-
-
-// - Event Collections - //
-/**
- * Named event collection.
- *
- * @param {string} name Name for this collection.
- * @param {boolean=} opt_skipb64encode Whether to skip base64 encoding. Pass as
- *        truthy when the collection name is already base64 encoded.
- * @constructor
- * @implements {bloombox.util.Exportable<proto.analytics.context.Collection>}
- * @implements {bloombox.util.Serializable}
- * @public
- */
-bloombox.telemetry.Collection = function Collection(name, opt_skipb64encode) {
-  /**
-   * Name for this collection.
-   *
-   * @type {string}
-   * @export
-   */
-  this.name = opt_skipb64encode ? name : bloombox.util.b64.encode(name);
-};
-
-
-/**
- * Static method to construct a `Collection` with an arbitrary string name.
- *
- * @param {string} name Name for this collection.
- * @return {bloombox.telemetry.Collection} Constructed collection.
- * @export
- */
-bloombox.telemetry.Collection.named = function(name) {
-  return new bloombox.telemetry.Collection(name);
-};
-
-
-/**
- * Export this `Collection` as an `analytics.Collection` message.
- *
- * @return {proto.analytics.context.Collection} JS PB message.
- */
-bloombox.telemetry.Collection.prototype.export = function() {
-  let collection = new proto.analytics.context.Collection();
-  collection.setName(this.name);
-  return collection;
-};
-
-
-/**
- * Render this collection object into a JSON-serializable structure suitable for
- * use over-the-wire.
- *
- * @return {Object}
- * @public
- */
-bloombox.telemetry.Collection.prototype.serialize = function() {
-  return {
-    'name': this.name
-  };
-};
 
 
 // - Master Context - //
@@ -652,327 +573,4 @@ bloombox.telemetry.Context.prototype.export = function() {
   if (this.browser) context.setBrowser(this.browser);
   if (this.native) context.setNative(this.native);
   return context;
-};
-
-
-// - Global Context - //
-/**
- * Globally-cached context singleton for values that are detected or loaded or
- * computed expensively by the runtime. Merged into event-level context before
- * events are sent.
- *
- * @type {?bloombox.telemetry.Context}
- * @package
- */
-bloombox.telemetry.GLOBAL_CONTEXT = null;
-
-
-/**
- * Globally-cached unique fingerprint for this device. Persisted in local
- * browser storage. Null means it is not yet initialized.
- *
- * @type {?string}
- * @package
- */
-bloombox.telemetry.DEVICE_FINGERPRINT = null;
-
-
-/**
- * Session-scoped token to indicate the border between different user sessions.
- * Leverages session storage instead of local storage. Null means it is not yet
- * initialized.
- *
- * @type {?string}
- * @package
- */
-bloombox.telemetry.SESSION_ID = null;
-
-
-/**
- * The key used in local storage to locally fingerprint a web browser-based
- * device.
- *
- * @const {string}
- * @package
- */
-bloombox.telemetry.DEVICE_FINGERPRINT_KEY = 'bb:v1:t.df';
-
-
-/**
- * The key used in session storage to indicate the session ID.
- *
- * @const {string}
- * @package
- */
-bloombox.telemetry.SESSION_ID_KEY = 'bb:v1:t.sid';
-
-
-/**
- * Resolve the device fingerprint, by creating it if it does not yet exist, or
- * returning the existing one if it does.
- *
- * @return {string} Global device fingerprint.
- */
-bloombox.telemetry.resolveFingerprint = function() {
-  if (bloombox.telemetry.DEVICE_FINGERPRINT === null) {
-    // try to fetch it from local storage
-    let existingFingerprint = (
-     window.localStorage.getItem(bloombox.telemetry.DEVICE_FINGERPRINT_KEY));
-
-    if (existingFingerprint && typeof existingFingerprint === 'string') {
-      // we found it, load the existing one from local storage
-      bloombox.telemetry.DEVICE_FINGERPRINT = existingFingerprint;
-    } else {
-      // we could not find one in local storage. generate one, persist it
-      // in local storage and locally, and return.
-      let newDeviceFingerprint = bloombox.util.generateUUID();
-      bloombox.telemetry.DEVICE_FINGERPRINT = newDeviceFingerprint;
-      window.localStorage.setItem(
-        bloombox.telemetry.DEVICE_FINGERPRINT_KEY, newDeviceFingerprint);
-      bloombox.logging.log('Established device fingerprint: "' +
-        newDeviceFingerprint + "'.");
-    }
-  }
-  return bloombox.telemetry.DEVICE_FINGERPRINT;
-};
-
-
-/**
- * Resolve the session ID, by creating it if it does not yet exist, or returning
- * the existing one if it does.
- *
- * @return {string} Session-scoped UUID.
- * @package
- */
-bloombox.telemetry.resolveSessionID = function() {
-  if (bloombox.telemetry.SESSION_ID === null) {
-    // try to fetch it from local storage
-    let existingID = (
-      window.sessionStorage.getItem(bloombox.telemetry.SESSION_ID_KEY));
-
-    if (existingID && typeof existingID === 'string') {
-      // we found it, load the existing one from local storage
-      bloombox.telemetry.SESSION_ID = existingID;
-    } else {
-      // we could not find one in local storage. generate one, persist it
-      // in local storage and locally, and return.
-      let newSessionID = bloombox.util.generateUUID();
-      bloombox.telemetry.SESSION_ID = newSessionID;
-      window.sessionStorage.setItem(
-        bloombox.telemetry.SESSION_ID_KEY, newSessionID);
-      bloombox.logging.log('Established user session at ID: "' +
-        newSessionID + "'.");
-    }
-  }
-  return bloombox.telemetry.SESSION_ID;
-};
-
-
-/**
- * Generate and return context for the current web application.
- *
- * @return {proto.analytics.context.WebApplication}
- */
-bloombox.telemetry.buildWebappContext = function() {
-  // collect data
-  let origin = window.location.origin;
-  let location = window.location.href;
-  let anchor = window.location.hash;
-  let protocol = window.location.protocol;
-  let title = document.title;
-  let referrer = document.referrer;
-
-  // fill out object
-  let webapp = new proto.analytics.context.WebApplication();
-  webapp.setOrigin(origin);
-  webapp.setLocation(location);
-  webapp.setTitle(title);
-  webapp.setProtocol(protocol);
-  webapp.setReferrer(referrer);
-  if (anchor) webapp.setAnchor(anchor);
-  return webapp;
-};
-
-
-/**
- * Build native device context from the available environment.
- *
- * @return {proto.analytics.context.NativeDeviceContext}
- * @package
- */
-bloombox.telemetry.buildNativeContext = function() {
-  let native = new proto.analytics.context.NativeDeviceContext();
-
-  // detect device type
-  let deviceType = proto.device.DeviceType.UNKNOWN_DEVICE_TYPE;
-  if (goog.labs.userAgent.device.isDesktop)
-    deviceType = proto.device.DeviceType.DESKTOP;
-  else if (goog.labs.userAgent.device.isTablet)
-    deviceType = proto.device.DeviceType.TABLET;
-  else if (goog.labs.userAgent.device.isMobile)
-    deviceType = proto.device.DeviceType.PHONE;
-  native.setType(deviceType);
-  native.setRole(proto.analytics.context.DeviceRole.CLIENT);
-
-  // touchpoints, concurrency and color depth
-  if (window.screen) {
-    // build screen information
-    let viewportHeight = window.screen.availHeight;
-    let viewportWidth = window.screen.availWidth;
-    let screenHeight = window.screen.height;
-    let screenWidth = window.screen.width;
-    let pixelDensity = window.devicePixelRatio;
-
-      let pixelSizeViewport = new proto.analytics.context.PixelSize();
-    pixelSizeViewport.setHeight(viewportHeight);
-    pixelSizeViewport.setWidth(viewportWidth);
-
-    let pixelSizeScreen = new proto.analytics.context.PixelSize();
-    pixelSizeScreen.setHeight(screenHeight);
-    pixelSizeScreen.setWidth(screenWidth);
-
-    let deviceScreen = new proto.analytics.context.DeviceScreen();
-    deviceScreen.setViewport(pixelSizeViewport);
-    deviceScreen.setScreen(pixelSizeScreen);
-    deviceScreen.setDensity(pixelDensity);
-    deviceScreen.setOrientation((window.innerHeight > window.innerWidth) ?
-      proto.analytics.context.ScreenOrientation.LANDSCAPE :
-      proto.analytics.context.ScreenOrientation.PORTRAIT);
-    native.setScreen(deviceScreen);
-  }
-
-  // detect OS type and version
-  let osType = proto.analytics.context.OSType.OS_UNKNOWN;
-  let osVersion = goog.userAgent.platform.VERSION;
-
-  if (goog.userAgent.IPAD ||
-    goog.userAgent.IPHONE ||
-    goog.userAgent.IPOD ||
-    goog.userAgent.IOS)
-    osType = proto.analytics.context.OSType.IOS;
-  else if ((goog.userAgent.WINDOWS ||
-      goog.userAgent.EDGE_OR_IE) &&
-    goog.userAgent.MOBILE)
-    osType = proto.analytics.context.OSType.WINDOWS_PHONE;
-  else if (goog.userAgent.WINDOWS ||
-    goog.userAgent.EDGE_OR_IE)
-    osType = proto.analytics.context.OSType.WINDOWS;
-  else if (goog.userAgent.product.ANDROID)
-    osType = proto.analytics.context.OSType.ANDROID;
-  else if (goog.userAgent.MAC)
-    osType = proto.analytics.context.OSType.MACOS;
-  else if (goog.userAgent.LINUX)
-    osType = proto.analytics.context.OSType.LINUX;
-
-  let osObj = new proto.analytics.context.DeviceOS();
-  let osVersionObj = new proto.structs.VersionSpec();
-  osVersionObj.setName(osVersion);
-  osObj.setType(osType);
-  osObj.setVersion(osVersionObj);
-  native.setOs(osObj);
-  return native;
-};
-
-
-/**
- * Build local browser context from the available environment.
- *
- * @return {proto.analytics.context.BrowserDeviceContext}
- * @package
- */
-bloombox.telemetry.buildBrowserContext = function() {
-  let context = new proto.analytics.context.BrowserDeviceContext();
-  let language = navigator.language;
-  let ua = navigator.userAgent;
-  context.setLanguage(language);
-  context.setUserAgent(ua);
-
-  if (typeof navigator.maxTouchPoints === 'number')
-    context.setTouchpoints(navigator.maxTouchPoints);
-  if (typeof navigator.hardwareConcurrency === 'number')
-    context.setHardwareConcurrency(navigator.hardwareConcurrency);
-  if (window.screen && typeof window.screen.colorDepth === 'number')
-    // set color depth
-    context.setColorDepth(window.screen.colorDepth);
-
-  // detect browser type and version
-  let browserVersion = goog.userAgent.VERSION;
-  let browserType = (
-    proto.analytics.context.BrowserType.BROWSER_UNKNOWN);
-  if (goog.userAgent.product.CHROME)
-    browserType = (
-      proto.analytics.context.BrowserType.CHROME);
-  else if (goog.userAgent.product.SAFARI)
-    browserType = (
-      proto.analytics.context.BrowserType.SAFARI);
-  else if (goog.userAgent.product.FIREFOX)
-    browserType = (
-      proto.analytics.context.BrowserType.FIREFOX);
-  else if (goog.userAgent.product.OPERA)
-    browserType = (
-      proto.analytics.context.BrowserType.OPERA);
-  else if (goog.userAgent.EDGE_OR_IE)
-    browserType = (
-      proto.analytics.context.BrowserType.IE_OR_EDGE);
-  context.setBrowserType(browserType);
-
-  // browser version mount
-  let browserVersionObj = new proto.structs.VersionSpec();
-  browserVersionObj.setName(browserVersion);
-  context.setVersion(browserVersionObj);
-  return context;
-};
-
-
-/**
- * Retrieve globally gathered/specified context. Caching is applied to reduce
- * overhead. To force a re-gather of expensively calculated information, pass
- * `opt_force_fresh` as truthy.
- *
- * @param {boolean=} opt_force_fresh Force a fresh load of global context.
- * @return {bloombox.telemetry.Context} Global context.
- * @public
- */
-bloombox.telemetry.globalContext = function(opt_force_fresh) {
-  let forceFresh = opt_force_fresh || false;
-  if (bloombox.telemetry.GLOBAL_CONTEXT === null || forceFresh) {
-    // grab global config
-    let config = bloombox.config.active();
-    let partnerCode = config.partner || null;
-    let locationCode = config.location || null;
-    let deviceFingerprint = bloombox.telemetry.resolveFingerprint();
-    let sessionID = bloombox.telemetry.resolveSessionID();
-    let nativeContext = bloombox.telemetry.buildNativeContext();
-    let browserContext = bloombox.telemetry.buildBrowserContext();
-
-    // build app context
-    let webContext = bloombox.telemetry.buildWebappContext();
-    let appContext = new proto.analytics.context.DeviceApplication();
-    appContext.setWeb(webContext);
-
-    if (bloombox.INTERNAL) {
-      // it's a bloombox app
-      appContext.setType(proto.analytics.context.ApplicationType.INTERNAL);
-    } else {
-      // it's a partner-side app
-      appContext.setType(proto.analytics.context.ApplicationType.PARTNER);
-    }
-
-    // calculate global context
-    bloombox.telemetry.GLOBAL_CONTEXT = new bloombox.telemetry.Context(
-      null,
-      partnerCode,
-      locationCode,
-      deviceFingerprint,
-      sessionID,
-      null,  // @TODO: ability to use logged-in user
-      null,  // @TODO: do we need a device here?
-      null,  // section key is set by callers
-      null,  // item key is set by callers
-      null,  // @TODO: ability to use active order
-      appContext,
-      browserContext,
-      nativeContext);
-  }
-  return bloombox.telemetry.GLOBAL_CONTEXT;
 };
